requirements.txt
# Backend dependencies
fastapi==0.115.12
python-dotenv==1.1.0
typing-extensions
pydantic==2.11.7
google-generativeai>=0.5.4
google.genai
uvicorn==0.23.2
python-multipart
python-docx>=1.1.0
Markdown>=3.6
beautifulsoup4>=4.12.2
html2text>=2020.1.16
markdown2>=2.4.10
docxtpl
docxcompose
pypandoc-binary

# Google Cloud Storage
google-cloud-storage>=2.10.0
google-cloud-aiplatform>=1.38.0

# Excel handling
openpyxl>=3.0.0
pycel==1.0b30

# Frontend dependencies
streamlit>=1.28.0
requests>=2.28.0
pandas>=1.5.0
httpx>=0.24.1
